Extract requirements diagram generation to a separate module
Clean up a bit by moving the functions responsible for generating the
requirements diagram properties into a separate module.

Also, update tests to account for changes.
